New List are now stored in the FireStore
I added a new state to represent the saving of data for displaying some
kind of loading indicator.
So far the only indication is, that the text on the button changes.
Also added code to send the entered data to the FireStore.

There are no checks on the entered data so far. This will come later.