Added event bus and first set of events
In this commit I added an EventBus. This event bus is responsible for
taking in events and notifying listeners about the hevents happening.
It does support listening to event streams, meaning events that happen
as a consequense of another event. This is essential, as one might
listen to a specific type of event, but only for the exact one that
happens as a consequence of a event just triggered.

An example would be the creation of a Course. In this case a
CourseCreated event will be triggered. Now we want to know when the new
course is persisted in the database. For that we have the
CoursePersisted event. So we would listen to this type of event. But
only listening to that type of event is not enough as we could by
mistake react to the CoursePersisted event of another course. Here the
idea of streams comes in. Streams are a way of connecting events
together, like a chain. A StreamListener enables you to listen to one
chain of events.