Added Aggregators to change the state as a reaction to events
Aggregators listen to certain types of events and update their state
according to the events they receive. They generate the state for the
query side of CQRS from the events triggered.

With this commit we have an general aggregator that stores all events as
well as an aggregator for Courses.