Differenciate between Events and Reactions
Events are things - more specifically Triggerables - that the user triggers. A direct consequence of his
interaction with the app.

A Reaction in contrast is Triggerable as well. But a Reaction can never
be a direct consequence of the users interaciton with the app. A
Reaction can olny be triggered through an preceeding Event.

Besides the technical definition, Events are the things we want to store
in our Event-Store to generate our state. While Reactions serve as
notifications that something has happened, which does not need to be
persisted.

As an example: Course Creation
The user creates a new course and hits the "add" button. This triggers
a CourseCreated Event - direct consequence of the users action. This
event is perisited in our Event-Store. We need it to
generate our state from it.
Here the CourseAggregator kicks in, it receives the CourseCreated Event
and updates it internal state, adding a new course to the course table.
When written, the CourseAggregator triggers the CoursePersisted
Reaction, to notify listeners about the course being now persisted in
the database. This event will finally get to our AddCourse Screen and
trigger the navigation to navigate back to the course overview.
We do not want to persist the CoursePersisted Event in our Event-Store
as it is not relevant for our state.