Real Authentication Flow
With this commit the App has a real authentication flow. For that we
wrapped the GoogleSignin logic in our own small interface, added
authentication flow ot the navigation and created the corresponding
views / screens for the authentication.

It now goes like this: App opens, InitAuth loads the authentication and
gets the initial state, meaning if the user is already logged in or not.
If the user is logged in, we directly redirect to the main App. If no
user is logge din yet, we display the SignIn screen.