Authentication with FireBase Backend
with this commit the credentials obtained from the Google sign in will
be used to authenticate with our firebase backend. This way we have our
Users persisted in our FireStore and also can access thier uid - which I
guess is what we want to use as a identifier in our Database.